share common realm code across Mac & iOS versions
specifically the code to set up the sync parts of Realm
and to create the initial list if necessary
